Close post-extraction dependency edges
