+++ crates/webcodex-core/src/lib.rs
@@ -12,6 +12,7 @@ pub mod job_observation;
 pub mod lsp_bridge;
 pub mod mcp_gateway;
 pub mod memory_contract;
+pub mod project_context_contract;
 pub mod project_instructions;
 pub mod project_listing;
 pub mod runner_protocol;
