+++ crates/webcodex-workspace/src/project_context.rs
@@ -15,6 +15,10 @@ use std::process::{Command, ExitStatus, Stdio};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, Instant, UNIX_EPOCH};
+pub use webcodex_core::project_context_contract::{
+    ContextFileFingerprint, FingerprintCompleteness, GitContextFingerprint,
+    ProjectContextFingerprint,
+};
 
 const FINGERPRINT_SCHEMA_VERSION: u32 = 2;
 pub const MAX_UNTRACKED_FILE_COUNT: usize = 512;
@@ -125,62 +129,6 @@ impl CaptureState {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ProjectContextFingerprint {
-    pub schema_version: u32,
-    /// Hash of the canonical absolute project path. The path itself is never
-    /// serialized into continuity state.
-    pub project_root_sha256: String,
-    pub target_directory: String,
-    pub git: GitContextFingerprint,
-    pub rules: Vec<ContextFileFingerprint>,
-    pub manifests: Vec<ContextFileFingerprint>,
-    #[serde(default)]
-    pub completeness: FingerprintCompleteness,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FingerprintCompleteness {
-    #[serde(default = "default_true")]
-    pub complete: bool,
-    #[serde(default)]
-    pub partial_slices: Vec<String>,
-    #[serde(default)]
-    pub warnings: Vec<String>,
-}
-
-impl Default for FingerprintCompleteness {
-    fn default() -> Self {
-        Self {
-            complete: true,
-            partial_slices: Vec::new(),
-            warnings: Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct GitContextFingerprint {
-    pub available: bool,
-    pub branch: Option<String>,
-    pub head: Option<String>,
-    pub worktree_sha256: Option<String>,
-    pub dirty: Option<bool>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ContextFileFingerprint {
-    pub path: String,
-    pub sha256: String,
-    pub bytes: u64,
-    #[serde(default = "default_true")]
-    pub complete: bool,
-    #[serde(default = "default_full_hash_kind")]
-    pub hash_kind: String,
-    #[serde(default)]
-    pub modified_unix_nanos: Option<u64>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ContextRefreshSummary {
     pub reused: Vec<String>,
     pub refreshed: Vec<String>,
@@ -201,14 +149,6 @@ pub struct ContextFileRefresh {
     pub removed: Vec<String>,
     #[serde(default)]
     pub unknown: Vec<String>,
-}
-
-fn default_true() -> bool {
-    true
-}
-
-fn default_full_hash_kind() -> String {
-    "full".to_string()
 }
 
 pub fn capture_project_context(
